削除対象を画像ファイルに限定

+++ min_file_delete_tool.php
@@ -2,8 +2,10 @@
 //globの引数に対象ディレクトリ名を指定
 foreach (glob('') as $file) {
     $size = filesize($file);
-    //glob配下にある20KB以下のファイルを削除
-    if (($size / 1024) < 20) {
+    //glob配下にある20KB以下の画像ファイルを削除
+    if (($size / 1024) < 20 &&
+        preg_match('/\.png$|\.jpg$|\.jpeg$/i', $file) === 1
+    ) {
         unlink($file);
     }
 }
